Add service for programmatic use, show notification immediately (#20)
* Add service for programmatic use, show notification immediately

This moves most work to a service, that emits an `update` event, so you could react to available updates also in different ways, like calling `window.location.reload()` automatically. The existing component is just another (default) consumer of that service.

This however implies a breaking change, in that passing a custom polling interval into the component does not work anymore, as the service is now responsible for polling. But AFAICT I don't see a use case for multiple component invocations by itself, especially with different polling intervals.

Fixes #19

* Rename service, fix broccoli transform

* Add FastBoot guard

* Extend docs with service usage