Guard against exceptions when calling navigator.serviceWorker.register()
Firefox seems to follow a policy that `navigator.serviceWorker.register()` is blocked when cookies are disabled. Due to that we see quite a few `SecurityError: The operation is insecure.` errors in our Sentry logs.

This change will just wrap the code in a try/catch, as I believe when the update function does not work for whatever reason, this should not be treated as an application-wide hard error, so rather be catched and not propagate up.